로그인, 로그아웃, 회원가입 로직 AuthController 으로 이동
