+++ recipe/src/main/resources/templates/index.html
@@ -167,8 +167,8 @@
 
   <!-- 오른쪽 메뉴 (로그인, 회원가입) -->
   <ul>
-    <li><a href="/api/members/login">로그인</a></li>
-    <li><a href="/api/members/register">회원가입</a></li>
+    <li><a href="/api/auth/login">로그인</a></li>
+    <li><a href="/api/auth/register">회원가입</a></li>
   </ul>
 </div>
