+++ recipe/src/main/resources/templates/login.html
@@ -72,7 +72,7 @@
 
 <div class="login-container">
     <h2>로그인</h2>
-    <form action="/api/members/login" method="POST">
+    <form action="/api/auth/login" method="POST">
         <div class="form-group">
             <label for="id">아이디</label>
             <input type="text" class="form-control" id="id" name="userID" placeholder="아이디" required>
@@ -87,7 +87,7 @@
 
     <!-- 회원가입 링크 -->
     <div class="signup-link">
-        <p>계정이 없으신가요? <a href="/api/members/register">회원가입</a></p>
+        <p>계정이 없으신가요? <a href="/api/auth/register">회원가입</a></p>
     </div>
 </div>
 </body>
