+++ recipe/src/main/resources/templates/register.html
@@ -74,7 +74,7 @@
 
 <div class="register-container">
   <h2>회원 가입</h2>
-  <form action="/api/members/register" method="POST">
+  <form action="/api/auth/register" method="POST">
 
     <div class="form-group">
       <label for="id">아이디</label>
